pkg/manifest: store the ostree container in the dir format on ISOs
The oci format cannot store signatures, leading to image builds
failing in the skopeo stage with:

Can not copy signatures to oci:/run/osbuild/tree/container:: Pushing signatures for OCI images is not supported
...
subprocess.CalledProcessError: Command '['skopeo', 'copy', 'containers-storage:[overlay@...', 'oci:/run/osbuild/tree/container']'

This commit thus switches the "intermediate" format to the dir one.
The dir format isn't stable, but since skopeo in the osbuild buildroot
and on the ISO should be of the same version, it's not a huge worry.